note-auto-poster/requirements.txt
anthropic
feedparser
requests
notion-client
python-dotenv
martian
